Outdated scripts removed, Parametric JobHandler bugs fixed, both Handlers flow improved, MasterScript changed to submit parametric jobs in batches of 50, scripts for cleaning up and downloading the blast and phyl. data added, test cases and example fasta files added, comments added

+++ BPM/SecondJobScript.sh
@@ -1,16 +1,25 @@
 #!/bin/bash -x
 
+#-- SecondJobScript.sh --------------------------------------------------------------------------------------
+#
+# This script makes phylogenetic profiles or runs mcl, according to "option" and "fileType" parameters
+# 
+# Arguments:
+# 1. The number of the uploaded abc files
+# 2. The name of the files I want to concatenate, "output" or "phyl"
+# 	"output" means simple abc file while "phyl" means abc file for phylogenetic profiles
+# 3. The User Name
+# 4. The selected program option
+# 5. Instance timestamp
+# 6. I_or_E for mcl
+#
+#------------------------------------------------------------------------------------------------------------
+
 date +"%Y%m%d %T"
 
 /bin/hostname -f
 lcg-cr --version
 
-# This script makes phylogenetic profiles or runs mcl, according to "option" and "fileType" parameters
-# first argument is the number of the uploaded abc files
-# second argument is the name of the files I want to concatenate "output" or "phyl"
-# "output" means simple abc file while "phyl" means abc file for phylogenetic profiles
-# third argument is the whoami where everything is uploaded
-# forth argument is the option the user has chosen
 
 abcNum=${1}
 fileType=${2}
@@ -24,13 +33,28 @@ export LFC_HOST=`lcg-infosites --vo see lfc`
 export LCG_CATALOG_TYPE=lfc
 export LCG_GFAL_VO=see
 
+################################################################################
+#                                                                              #
+#                     Download files and MCL                                   #
+#                                                                              #
+################################################################################
+
 echo "abc num: $abcNum"
 echo "file type: $fileType"
 # get the abc files and concatenate them
-if [[ $fileType != "phyl" || $option != 5 ]]; then
+#if [[ $fileType != "phyl" || $option != 5 ]]; then
 	for i in `seq 1 $abcNum`
 	do
 		lcg-cp lfn:/grid/see/$userName/BPM_$timestamp/$fileType$i.tar.gz file:$fileType$i.tar.gz
+		if [[ $? != 0 ]]; then
+			lcg-ls lfn:/grid/see/$userName/BPM_$timestamp/$fileType$i.tar.gz
+			if [[ $? == 0 ]]; then
+			# its there, try again
+			lcg-cp lfn:/grid/see/$userName/BPM_$timestamp/$fileType$i.tar.gz file:$fileType$i.tar.gz
+			else
+				echo "$i" > failed
+			fi
+		fi
 		# untar them
 		tar -zxvf $fileType$i.tar.gz
 		rm $fileType$i.tar.gz
@@ -40,7 +64,8 @@ if [[ $fileType != "phyl" || $option != 5 ]]; then
 	echo "$abcNum"
 	echo "$untaredNum"
 	if [[ $untaredNum != $abcNum ]]; then
-		echo "didn't download correctly all the files"
+		echo "didn't download correctly all the files, some failed:"
+		cat failed
 		exit
 	else
 		echo "cat and maybe sort"
@@ -53,7 +78,7 @@ if [[ $fileType != "phyl" || $option != 5 ]]; then
 		fi
 		echo "done cat"
 		date +"%Y%m%d %T"
-		if [[ $fileType == "output" || $option == 3 || $option == 4 ]]; then # option 5 has mcl of simple output but not of phyl profile
+#		if [[ $fileType == "output" || $option == 3 || $option == 4 ]]; then # option 5 has mcl of simple output but not of phyl profile
 			# get and install mcl
 			wget http://micans.org/mcl/src/mcl-latest.tar.gz
 		
@@ -77,17 +102,32 @@ if [[ $fileType != "phyl" || $option != 5 ]]; then
 			cp $HOME/bin/mcl .
 			cp $HOME/bin/mcxload .
 			
-			if [[ $I_or_E == "E" ]]; then
+			if [[ $I_or_E == "E" && $fileType == "output" ]]; then
 				# run mcl
-				./mcxload -abc Output.abc --stream-neg-log10 -o out.mci -write-tab out.tab
-				./mcl out.mci -use-tab out.tab -o Out$fileType.mcl
+				time ./mcxload -abc Output.abc --stream-mirror --stream-neg-log10 -o out.mci --write-binary
+				time ./mcl out.mci -use-tab out.tab -o tempMcl 
 			else
-				time ./mcl Output.abc --abc -o Out$fileType.mcl
+                        	#time ./mcxload -abc Output.abc -o out.mci --write-binary
+				#time ./mcl out.mci -o tempMcl 
+				#time ./mcl Output.abc --abc -o tempMcl
+				time ./mcxload -abc Output.abc --stream-mirror -o out.mci -write-tab out.tab --write-binary
+				time ./mcl out.mci -use-tab out.tab -o tempMcl
 			fi
+			# remove the clusters with only one element
+			awk '{if(NF >=2 ){print}}' tempMcl > tempMcl2	
+			# insert arithmetic id for every cluster
+			awk '{print ">"NR,$0}' tempMcl2 > Out$fileType.mcl
+			
 			date +"%Y%m%d %T"
 		fi
-	fi
-fi
+#	fi
+#fi
+
+################################################################################
+#                                                                              #
+#           Download files and complete Phylogenetic Profiles                  #
+#                                                                              #
+################################################################################
 
 if [[ $fileType == "phyl" ]]; then
 	if [[ $option == 5 ]]; then
@@ -97,6 +137,16 @@ if [[ $fileType == "phyl" ]]; then
 		for i in `seq 1 $abcNum`
 		do
 		        lcg-cp lfn:/grid/see/$userName/BPM_$timestamp/PhylProf$i.tar.gz file:PhylProf$i.tar.gz
+			if [[ $? != 0 ]]; then
+                        	lcg-ls lfn:/grid/see/$userName/BPM_$timestamp/PhylProf$i.tar.gz
+                        	if [[ $? == 0 ]]; then
+                        		# its there, try again
+                           		lcg-cp lfn:/grid/see/$userName/BPM_$timestamp/PhylProf$i.tar.gz file:PhylProf$i.tar.gz
+                        	else
+                                	echo "$i" >> failed
+                        	fi
+                	fi
+
 		        # untar them
 		        tar -zxvf PhylProf$i.tar.gz
 		        rm PhylProf$i.tar.gz
@@ -107,7 +157,7 @@ if [[ $fileType == "phyl" ]]; then
 		if [[ $untaredNum != $abcNum ]]; then	
 			echo "could not get all the phylogenetic profiles"
 		else
-			cat tmp PhylProf* > PhylogeneticProfile
+			cat PhylProf* > PhylogeneticProfile.txt
 			echo "done making phyl prof"
 		fi
 		echo "done making phyl prof"
@@ -121,10 +171,10 @@ if [[ $fileType == "phyl" ]]; then
 		rm tmp
 		t=$(grep -m 1 -f querySequences Output.abc | cut -f 1)
 		grep "$t" Output.abc | cut -f 2 | sed 's/^//' | tr '\n' ' ' > genomes
-		echo -e "`cat genomes`\n" > PhylogeneticProfile
+		echo -e "`cat genomes`\n" > PhylogeneticProfile.txt
 		while read line; do
-			grep "^$line" Output.abc | cut -f 3 | sed 's/^//' | tr '\n' ' ' > tmp
-      			echo "$line `cat tmp`"  >> PhylogeneticProfile
+			grep -w "^$line" Output.abc | cut -f 3 | sed 's/^//' | tr '\n' ' ' > tmp
+      			echo "$line `cat tmp`"  >> PhylogeneticProfile.txt
        			rm tmp
 		done < querySequences
 		echo "done making phyl prof"
@@ -132,20 +182,26 @@ if [[ $fileType == "phyl" ]]; then
 	fi
 fi
 
+################################################################################
+#                                                                              #
+#                           Upload output                                      #
+#                                                                              #
+################################################################################
+
 chmod +x Upload.sh
 if [[ $fileType == "output" ]]; then
 	mv Out$fileType.mcl Outsimple.mcl
-	tar -zcvf resultssimple.tar.gz Outsimple.mcl
-	./Upload.sh resultssimple.tar.gz $userName $timestamp
+	tar -zcvf resultssimpleF.tar.gz Outsimple.mcl
+	./Upload.sh resultssimpleF.tar.gz $userName $timestamp
 else	
 	case "$option" in
 	
-	2|5) tar -zcvf resultsphyl.tar.gz PhylogeneticProfile
-	     ./Upload.sh resultsphyl.tar.gz $userName $timestamp
+	2|5) tar -zcvf resultsphylF.tar.gz PhylogeneticProfile.txt
+	     ./Upload.sh resultsphylF.tar.gz $userName $timestamp
 	   ;;
 	3|4) mv Out$fileType.mcl Outphyl.mcl
-	     tar -zcvf resultsphyl.tar.gz Outphyl.mcl PhylogeneticProfile
-	     ./Upload.sh resultsphyl.tar.gz $userName $timestamp
+	     tar -zcvf resultsphylF.tar.gz Outphyl.mcl PhylogeneticProfile.txt
+	     ./Upload.sh resultsphylF.tar.gz $userName $timestamp
 	   ;;
 	esac	
 	echo "done"
